complete expression calculator

+++ data_structure/data_structure/data_structure/expression_calculator.h
@@ -1,8 +1,5 @@
 #pragma once
 
 
-uint8 calculate(String infix_expression, char * suffix_expression);
-
-
-
+double calculate(String infix_expression, char * suffix_expression);
